--- application/DelayProvider/DelayProvider.hpp
@@ -1,10 +1,0 @@
-#pragma once
-
-#include <stdint.h>
-
-namespace ATC {
-class DelayProvider {
-public:
-    void delayMiliseconds(const uint32_t value);
-};
-}
