+++ application/Hardware/FramebufferDisplay/Ws17143Display/Ws17143Display.cpp
@@ -491,7 +491,7 @@ void Ws17143Display::displayFramebuffer() {
 Ws17143Display::Ws17143Display(
     const Ws17143DisplayPinout& pinout,
     FlexibleMemoryController& flexibleMemoryController,
-    DelayProvider& delayProvider
+    IDelayProvider& delayProvider
 ) :
     pinout_(pinout),
     flexibleMemoryController_(flexibleMemoryController),
