+++ atc-nucleo-u5a5zj/Adapters/DelayProvider/DelayProvider.cpp
@@ -1,13 +1,9 @@
 #include "DelayProvider.hpp"
 
-#include <stdint.h>
-
-// TODO remove this dependency
-#include "../../atc-nucleo-u5a5zj/Core/Inc/main.h"
+#include "main.h"
 
 namespace ATC {
 void DelayProvider::delayMiliseconds(const uint32_t value) {
-    // TODO abstract it out
     HAL_Delay(value);
 }
 }
