+++ atc-nucleo-u5a5zj/Adapters/DelayProvider/DelayProvider.hpp
@@ -1,0 +1,10 @@
+#pragma once
+
+#include "application/Ports/IDelayProvider.hpp"
+
+namespace ATC {
+class DelayProvider : public IDelayProvider {
+public:
+    virtual void delayMiliseconds(const uint32_t value) override;
+};
+}
